@@ -34,7 +34,7 @@
     "event-emitter": "^0.3.4",
     "foundation-apps": "git+https://github.com/zurb/foundation-apps.git",
     "git-rev-sync": "^1.4.0",
-    "graphenejs-lib": "^0.4.3",
+    "graphenejs-lib": "^0.4.6",
     "graphenejs-ws": "^0.1.3",
     "highcharts": "^4.2.3",
     "history": "1.13.1",
@@ -50,11 +50,11 @@
     "object-assign": "^4.0.1",
     "peerjs": "^0.3.14",
     "perfect-scrollbar": "^0.6.11",
-    "react": "^15.1.0",
+    "react": "^15.2.0",
     "react-addons-css-transition-group": "^15.1.0",
-    "react-dom": "^15.1.0",
+    "react-dom": "^15.2.0",
     "react-foundation-apps": "git+https://github.com/valzav/react-foundation-apps.git",
-    "react-highcharts": "^8.3.0",
+    "react-highcharts": "^8.4.1",
     "react-interpolate-component": "^0.9.0",
     "react-intl": "^2.0.1",
     "react-json-inspector": "^7.0.0",
@@ -63,7 +63,7 @@
     "react-popover": "^0.4.4",
     "react-responsive-mixin": "^0.3.5",
     "react-router": "1.0.1",
-    "react-tooltip": "^3.0.0",
+    "react-tooltip": "^3.0.8",
     "react-translate-component": "^0.11.0",
     "svgo": "^0.5.1",
     "tcomb": "2.5.2",
