@@ -6,7 +6,7 @@ import AccountStore from "stores/AccountStore";
 import AccountNameInput from "./../Forms/AccountNameInput";
 import PasswordInput from "./../Forms/PasswordInput";
 import WalletDb from "stores/WalletDb";
-import notify from 'actions/NotificationActions';
+import notify from "actions/NotificationActions";
 import {Link} from "react-router";
 import AccountImage from "./AccountImage";
 import AccountSelect from "../Forms/AccountSelect";
@@ -16,18 +16,18 @@ import LoadingIndicator from "../LoadingIndicator";
 import WalletActions from "actions/WalletActions";
 import Translate from "react-translate-component";
 import RefcodeInput from "../Forms/RefcodeInput";
-import {TransitionMotion, spring} from 'react-motion';
 import {ChainStore, FetchChain} from "graphenejs-lib";
+import {BackupCreate} from "../Wallet/Backup";
 
 @connectToStores
 class CreateAccount extends React.Component {
 
     static getStores() {
-        return [AccountStore]
+        return [AccountStore];
     };
 
     static getPropsFromStores() {
-        return {}
+        return {};
     };
 
     constructor() {
@@ -39,7 +39,8 @@ class CreateAccount extends React.Component {
             registrar_account: null,
             loading: false,
             hide_refcode: true,
-            show_identicon: false
+            show_identicon: false,
+            step: 1
         };
         this.onFinishConfirm = this.onFinishConfirm.bind(this);
 
@@ -53,7 +54,8 @@ class CreateAccount extends React.Component {
             nextState.registrar_account !== this.state.registrar_account ||
             nextState.loading !== this.state.loading ||
             nextState.hide_refcode !== this.state.hide_refcode ||
-            nextState.show_identicon !== this.state.show_identicon;
+            nextState.show_identicon !== this.state.show_identicon ||
+            nextState.step !== this.state.step;
     }
 
     isValid() {
@@ -83,7 +85,7 @@ class CreateAccount extends React.Component {
 
             FetchChain("getAccount", this.state.accountName).then(() => {
                 console.log("onFinishConfirm");
-                this.props.history.pushState(null, `/wallet/backup/create?newAccount=true`);    
+                this.props.history.pushState(null, `/wallet/backup/create?newAccount=true`);
             });
         }
     }
@@ -99,9 +101,12 @@ class CreateAccount extends React.Component {
                     TransactionConfirmStore.listen(this.onFinishConfirm);
                 } else { // Account registered by the faucet
                     console.log("account registed by faucet");
-                    this.props.history.pushState(null, `/wallet/backup/create?newAccount=true`);
+                    // this.props.history.pushState(null, `/wallet/backup/create?newAccount=true`);
+                    this.setState({
+                        step: 2
+                    });
                     // this.props.history.pushState(null, `/account/${name}/overview`);
-                    
+
                 }
             }).catch(error => {
                 console.log("ERROR AccountActions.createAccount", error);
@@ -154,147 +159,231 @@ class CreateAccount extends React.Component {
         this.setState({hide_refcode: false});
     }
 
-    getHeaderItemStyles() {
-        let config = {};
-        let d = {
-            opacity: spring(1),
-            top: spring(16)
-        };
-        if (this.state.show_identicon) config["icon"] = d;
-        else config["title"] = d;
-        return config;
-    }
+    _renderAccountCreateForm() {
 
-    headerItemWillEnter(key) {
-        return {
-            opacity: spring(0),
-            top: spring(-50)
-        };
-    }
-
-    headerItemWillLeave(key) {
-        return {
-            opacity: spring(0),
-            top: spring(-50)
-        };
-    }
-
-    render() {
         let {registrar_account} = this.state;
 
-        let my_accounts = AccountStore.getMyAccounts()
+        let my_accounts = AccountStore.getMyAccounts();
         let firstAccount = my_accounts.length === 0;
-
         let hasWallet = WalletDb.getWallet();
         let valid = this.isValid();
 
         let isLTM = false;
-        let registrar = registrar_account ? ChainStore.getAccount(registrar_account) : null; 
+        let registrar = registrar_account ? ChainStore.getAccount(registrar_account) : null;
         if (registrar) {
-            if( registrar.get( 'lifetime_referrer' ) == registrar.get( 'id' ) ) {
+            if( registrar.get( "lifetime_referrer" ) == registrar.get( "id" ) ) {
                 isLTM = true;
             }
         }
 
         let buttonClass = classNames("button no-margin", {disabled: (!valid || (registrar_account && !isLTM))});
 
-        let header_items = {
-            icon: <div className="form-group">
-                <label><Translate content="account.identicon"/></label>
-                <AccountImage account={this.state.validAccountName ? this.state.accountName : null}/>
-            </div>,
-            title: firstAccount ?
-                    (<div>
-                        <h1><Translate content="account.welcome"/></h1>
-                        <h3><Translate content="account.please_create_account"/></h3>
-                        <hr/>
-                    </div>) :
-                    (
-                        <div>
-                            <h1><Translate content="account.create_account"/></h1>
-                            <hr/>
-                        </div>
-                    )
-        };
+        return (
+            <form
+                style={{maxWidth: "40rem"}}
+                onSubmit={this.onSubmit.bind(this)}
+                noValidate
+            >
+                <AccountNameInput
+                    ref={(ref) => {if (ref) {this.accountNameInput = ref.refs.nameInput;}}}
+                    cheapNameOnly={firstAccount}
+                    onChange={this.onAccountNameChange.bind(this)}
+                    accountShouldNotExist={true}
+                    placeholder="Account Name (Public)"
+                    noLabel
+                />
 
-        const header = <TransitionMotion
-            styles={this.getHeaderItemStyles()}
-            willEnter={this.headerItemWillEnter}
-            willLeave={this.headerItemWillLeave}>
-            {config =>
-                <div>
-                    {Object.keys(config).map(key =>
-                        {
-                            let style = config[key];
-                            return <div key={key} style={{position: "absolute", left: 0, right: 0, ...style}}>
-                                <div className="center-content">{header_items[key]}</div>
-                            </div>;
-                        })
-                    }
+                {/* Only ask for password if a wallet already exists */}
+                {hasWallet ?
+                    null :
+
+                        <PasswordInput
+                            ref="password"
+                            confirmation={true}
+                            onChange={this.onPasswordChange.bind(this)}
+                            noLabel
+                        />
+                }
+
+                {/* If this is not the first account, show dropdown for fee payment account */}
+                {
+                firstAccount ? null : (
+                    <div className="full-width-content form-group no-overflow">
+                        <label><Translate content="account.pay_from" /></label>
+                        <AccountSelect
+                            account_names={my_accounts}
+                            onChange={this.onRegistrarAccountChange.bind(this)}
+
+                        />
+                        {(registrar_account && !isLTM) ? <div style={{textAlign: "left"}} className="facolor-error"><Translate content="wallet.must_be_ltm" /></div> : null}
+                    </div>)
+                }
+
+                {/* Submit button */}
+                {this.state.loading ?  <LoadingIndicator type="three-bounce"/> : <button className={buttonClass}><Translate content="account.create_account" /></button>}
+
+                {/* Backup restore option */}
+                <div style={{paddingTop: 40}}>
+                    <label style={{textTransform: "none"}}>
+                        <Link to="/existing-account">
+                            <Translate content="wallet.restore" />
+                        </Link>
+                    </label>
                 </div>
-            }
-        </TransitionMotion>
+
+                {/* Skip to step 3 */}
+                {(!hasWallet || firstAccount ) ? null :<div style={{paddingTop: 20}}>
+                    <label style={{textTransform: "none"}}>
+                        <a onClick={() => {this.setState({step: 3});}}><Translate content="wallet.go_get_started" /></a>
+                    </label>
+                </div>}
+            </form>
+        );
+    }
+
+    _renderAccountCreateText() {
+        let hasWallet = WalletDb.getWallet();
+        let my_accounts = AccountStore.getMyAccounts();
+        let firstAccount = my_accounts.length === 0;
 
         return (
-            <div className="grid-container">
-                <div className="create-account-header">
-                    {header}
+            <div>
+                <p style={{fontWeight: "bold"}}><Translate content="wallet.wallet_browser" /></p>
+
+                <p>{!hasWallet ? <Translate content="wallet.has_wallet" /> : null}</p>
+
+                <Translate style={{textAlign: "left"}} component="p" content="wallet.create_account_text" />
+
+                {firstAccount ? <Translate style={{textAlign: "left"}} component="p" content="wallet.first_account_paid" /> : null}
+
+                {this.state.hide_refcode ? null :
+                    <div>
+                        <RefcodeInput ref="refcode" label="refcode.refcode_optional" expandable={true}/>
+                        <br/>
+                    </div>
+                }
+            </div>
+        );
+    }
+
+    _renderBackup() {
+        return (
+            <div>
+                <BackupCreate noText downloadCb={this._onBackupDownload}/>
+            </div>
+        );
+    }
+
+    _onBackupDownload = () => {
+        this.setState({
+            step: 3
+        });
+    }
+
+    _renderBackupText() {
+        return (
+            <div>
+                <p style={{fontWeight: "bold"}}><Translate content="footer.backup" /></p>
+                <p><Translate content="wallet.wallet_crucial" /></p>
+                <p><Translate content="wallet.wallet_move" /></p>
+                <p><Translate content="wallet.wallet_lose_warning" /></p>
+            </div>
+        );
+    }
+
+    _renderGetStarted() {
+
+        return (
+            <div>
+                <table className="table">
+                    <tbody>
+
+                        <tr>
+                            <td><Translate content="wallet.tips_dashboard" />:</td>
+                            <td><Link to="dashboard"><Translate content="header.dashboard" /></Link></td>
+                        </tr>
+
+                        <tr>
+                            <td><Translate content="wallet.tips_account" />:</td>
+                            <td><Link to={`/account/${this.state.accountName}/overview`} ><Translate content="wallet.link_account" /></Link></td>
+                        </tr>
+
+                        <tr>
+                            <td><Translate content="wallet.tips_deposit" />:</td>
+                            <td><Link to="deposit-withdraw"><Translate content="wallet.link_deposit" /></Link></td>
+                        </tr>
+
+
+
+                        <tr>
+                            <td><Translate content="wallet.tips_transfer" />:</td>
+                            <td><Link to="transfer"><Translate content="wallet.link_transfer" /></Link></td>
+                        </tr>
+
+                        <tr>
+                            <td><Translate content="wallet.tips_settings" />:</td>
+                            <td><Link to="settings"><Translate content="header.settings" /></Link></td>
+                        </tr>
+                    </tbody>
+
+                </table>
+            </div>
+        );
+    }
+
+    _renderGetStartedText() {
+
+        return (
+            <div>
+                <p style={{fontWeight: "bold"}}><Translate content="wallet.congrat" /></p>
+
+                <p><Translate content="wallet.tips_explore" /></p>
+
+                <p><Translate content="wallet.tips_header" /></p>
+
+                <p style={{fontWeight: "bold"}}><Translate content="wallet.tips_login" /></p>
+            </div>
+        );
+    }
+
+    render() {
+        let {step} = this.state;
+
+        let my_accounts = AccountStore.getMyAccounts();
+        let firstAccount = my_accounts.length === 0;
+
+        return (
+            <div className="grid-block vertical page-layout">
+                <div className="grid-container shrink">
+                    <div style={{textAlign: "center", paddingTop: 20}}>
+                        <Translate content="wallet.wallet_new" component="h2" />
+
+                        <h4 style={{paddingTop: 20}}>
+                            {step === 1 ?
+                                <span>{firstAccount ? <Translate content="wallet.create_w_a" />  : <Translate content="wallet.create_a" />}</span> :
+                            step === 2 ? <Translate content="wallet.create_success" /> :
+                            <Translate content="wallet.all_set" />
+                        }
+                        </h4>
+                    </div>
                 </div>
-                <div className="content-block center-content">
-                        <form
-                            style={{maxWidth: "45rem"}}
-                            onSubmit={this.onSubmit.bind(this)}
-                            noValidate
-                        >
-                            <div className="grid-content">
-                                <Translate style={{textAlign: "left"}} component="p" content="wallet.create_account_text" />
-                                {firstAccount ? <Translate style={{textAlign: "left"}} component="p" content="wallet.first_account_paid" /> : null}
-                                <AccountNameInput
-                                    ref={(ref) => {if (ref) {this.accountNameInput = ref.refs.nameInput;}}}
-                                    cheapNameOnly={firstAccount}
-                                    onChange={this.onAccountNameChange.bind(this)}
-                                    accountShouldNotExist={true}
-                                />
-                            </div>
-                            {hasWallet ?
-                                null :
-                                <div className="grid-content">
-                                    <div style={{
-                                        textAlign: "left",
-                                        paddingBottom: 10
-                                    }}>
-                                        <Translate component="p" content="wallet.account_create_wallet_text" />
-                                    </div>
-                                    <PasswordInput ref="password" confirmation={true} onChange={this.onPasswordChange.bind(this)}/>
-                                </div>
-                            }
-                            {
-                            firstAccount ? null : (
-                                <div className="full-width-content grid-content form-group no-overflow">
-                                    <label><Translate content="account.pay_from" /></label>
-                                    <AccountSelect
-                                        account_names={my_accounts}
-                                        onChange={this.onRegistrarAccountChange.bind(this)}
-                                    />
-                                    {(registrar_account && !isLTM) ? <div style={{textAlign: "left"}} className="facolor-error"><Translate content="wallet.must_be_ltm" /></div> : null}
-                                </div>)
-                            }
-                            {this.state.hide_refcode ? null :
-                                <div>
-                                    <RefcodeInput ref="refcode" label="refcode.refcode_optional" expandable={true}/>
-                                    <br/>
-                                </div>
-                            }
-                            {this.state.loading ?  <LoadingIndicator type="three-bounce"/> : <button className={buttonClass}><Translate content="account.create_account" /></button>}
-                            
-                            <div style={{paddingTop: 20}}>
-                                <label className="inline">
-                                    <Link to="/existing-account">
-                                        <Translate content="wallet.restore" />
-                                    </Link>
-                                </label>
-                            </div>
-                        </form>
+                <div className="grid-block main-content wrap" style={{marginTop: "2rem"}}>
+                    <div className="grid-content small-12 medium-6" style={{paddingLeft: "15%"}}>
+                        <p style={{fontWeight: "bold"}}>
+                            <Translate content={"wallet.step_" + step} />
+                        </p>
+
+                        {step === 1 ? this._renderAccountCreateForm() : step === 2 ? this._renderBackup() :
+                            this._renderGetStarted()
+                        }
+                    </div>
+
+                    <div className="grid-content small-12 medium-6" style={{paddingRight: "15%"}}>
+                        {step === 1 ? this._renderAccountCreateText() : step === 2 ? this._renderBackupText() :
+                            this._renderGetStartedText()
+                        }
+
+                    </div>
                 </div>
             </div>
         );
