@@ -32,7 +32,7 @@ class Settings extends React.Component {
             ],
             settingEntries: {
                 general: ["locale", "unit", "showSettles", "walletLockTimeout", "themes",
-                "disableChat"],
+                "disableChat", "showAssetPercent"],
                 access: ["connection", "faucet_address"]
             }
         };
@@ -102,6 +102,10 @@ class Settings extends React.Component {
 
         case "showSettles":
             SettingsActions.changeSetting({setting: "showSettles", value: e.target.value === "yes" });
+            break;
+
+        case "showAssetPercent":
+            SettingsActions.changeSetting({setting: "showAssetPercent", value: e.target.value === "yes" });
             break;
 
         case "unit":
