@@ -133,9 +133,9 @@ class TotalValue extends React.Component {
 
     _assetValues(totals, amount, asset) {
         if (!totals[asset]) {
-            totals[asset] = amount; 
+            totals[asset] = amount;
         } else {
-            totals[asset] += amount; 
+            totals[asset] += amount;
         }
 
         return totals;
@@ -144,7 +144,7 @@ class TotalValue extends React.Component {
     render() {
         let {fromAssets, toAsset, balances, marketStats, collateral, debt, openOrders, inHeader} = this.props;
         let coreAsset = ChainStore.getAsset("1.3.0");
-        
+
         if (!coreAsset || !toAsset) {
             return null;
         }
@@ -246,13 +246,13 @@ class TotalValue extends React.Component {
             totalsTip += `<tr><td>&nbsp;</td><td style="text-align: right;">${noDataSymbol} no data</td></tr>`;
 
         totalsTip += "</tbody></table>"
-        
+
         // console.log("assetValues:", assetValues, "totalsTip:", totalsTip);
         if (!inHeader) {
             return <FormattedAsset amount={totalValue} asset={toAsset.get("id")} decimalOffset={toAsset.get("symbol").indexOf("BTC") === -1 ? toAsset.get("precision") : 4}/>;
         } else {
             return (
-                <div data-tip={totalsTip} data-place="bottom" data-type="light" html data-html={true} >
+                <div data-tip={totalsTip} data-place="bottom" data-type="light" data-html={true} >
                     <FormattedAsset amount={totalValue} asset={toAsset.get("id")} decimalOffset={toAsset.get("symbol").indexOf("BTC") === -1 ? toAsset.get("precision") : 4}/>
                 </div>
 
